Machine Learning with Cloud Computing/Practical 1/requirements.txt
pandas==2.2.0
pyarrow==15.0.0 ## Pyarrow will become a required dependency of pandas in the next major release of pandas (pandas 3.0)
python-dotenv==1.0.1
